// src/types/graphql-generated/graphql.d.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type Ticket = {
  __typename?: 'Ticket';
  comment: Maybe<Scalars['String']>;
  createdAt: Maybe<Scalars['String']>;
  description: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  priority: TicketPriority;
  projectId: Scalars['String'];
  references: Maybe<Array<Maybe<Scalars['String']>>>;
  status: TicketStatus;
  title: Scalars['String'];
  updatedAt: Maybe<Scalars['String']>;
  userId: Scalars['String'];
};
// ...
export enum TicketPriority {
  HIGHT = 'HIGHT',
  LOW = 'LOW',
  MEDIUM = 'MEDIUM',
  URGENT = 'URGENT'
}

export enum TicketStatus {
  ARCHIVED = 'ARCHIVED',
  BACKLOG = 'BACKLOG',
  BLOCKED = 'BLOCKED',
  DONE = 'DONE',
  IN_PROGRESS = 'IN_PROGRESS',
  REVIEW = 'REVIEW',
  TO_DO = 'TO_DO'
}

export type TicketUpdateInput = {
  comment: InputMaybe<Scalars['String']>;
  description: InputMaybe<Scalars['String']>;
  priority: InputMaybe<TicketPriority>;
  projectId: InputMaybe<Scalars['String']>;
  references: InputMaybe<Array<InputMaybe<Scalars['String']>>>;
  status: InputMaybe<TicketStatus>;
  ticketId: Scalars['ID'];
  title: InputMaybe<Scalars['String']>;
};

export type TicketsPayload = {
  __typename?: 'TicketsPayload';
  tickets: Maybe<Array<Ticket>>;
  userErrors: Array<UserError>;
};

export type User = {
  __typename?: 'User';
  confirmed: Maybe<Scalars['Boolean']>;
  createdAt: Maybe<Scalars['String']>;
  email: Scalars['String'];
  id: Scalars['ID'];
  name: Scalars['String'];
  projects: Maybe<Array<Project>>;
  tickets: Maybe<Array<Ticket>>;
  updatedAt: Maybe<Scalars['String']>;
};

export type UserError = {
  __typename?: 'UserError';
  message: Scalars['String'];
  values: Maybe<Array<Maybe<Scalars['String']>>>;
};